Add support for a range of child process ports
Allows the user to specify --child-process-ports MIN-MAX to specify the
range of ports to use for fork'd/exec'd processes rather than picking a
random port from the ephemeral port range

This is useful when you cannot open up all the ports in that region very
easily (i.e. docker containers).